Bolši selectbox za izbor oseb in predmetov v kombiniranem pogledu.
