+++ urnik/models.py
@@ -3,12 +3,7 @@ from collections import defaultdict
 from copy import deepcopy
 
 from django.conf import settings
-from django.core.exceptions import ValidationError
 from django.db import models
-from django.forms import ModelForm, CheckboxSelectMultiple, TextInput, DateInput, BooleanField, ModelMultipleChoiceField
-from django.template import defaultfilters
-
-from urnik.templatetags.tags import dan_tozilnik_mnozina
 from urnik.utils import teden_dneva
 from .layout import nastavi_sirine_srecanj, nastavi_barve
 
@@ -161,7 +156,7 @@ class Ucilnica(models.Model):
 
     def __str__(self):
         return self.oznaka
-    
+
     def __lt__(self, other):
         return self.oznaka < other.oznaka
 
@@ -186,6 +181,13 @@ class Predmet(models.Model):
     def __str__(self):
         return self.ime
 
+    def opisno_ime(self):
+        letniki = self.kratice_letnikov()
+        if not letniki:
+            return self.ime
+        else:
+            return "{} ({})".format(self.ime, letniki)
+
     def kratice_letnikov(self):
         return ', '.join(letnik.kratica for letnik in self.letniki.all())
 
@@ -209,7 +211,7 @@ class Semester(models.Model):
 
     def __str__(self):
         return self.ime
-    
+
     def pripravi_kopijo(self):
         novi_semester = Semester.objects.create(
             ime=self.ime + ' (kopija)',
@@ -571,127 +573,3 @@ class Rezervacija(models.Model):
     def se_po_urah_prekriva(self, od, do):
         return self.od < do and od < self.do
 
-
-class RezervacijeModelMultipleChoiceField(ModelMultipleChoiceField):
-    def __init__(self, *args, **kwargs):
-        ModelMultipleChoiceField.__init__(self, *args, **kwargs)
-        it = self.iterator(self)
-        self.choices = [it.choice(obj) for obj in self.queryset]
-
-
-class OsebeModelMultipleChoiceField(RezervacijeModelMultipleChoiceField):
-    def label_from_instance(self, obj):
-        return obj.priimek_ime
-
-
-class UcilniceModelMultipleChoiceField(RezervacijeModelMultipleChoiceField):
-    def label_from_instance(self, obj):
-        return obj.oznaka
-
-
-class RezevacijeForm(ModelForm):
-
-    PREKRIVANJA = 'prekrivanja'
-
-    osebe = OsebeModelMultipleChoiceField(queryset=Oseba.objects.all(),
-                                          help_text='Osebe, ki si lastijo to rezervacijo')
-    ucilnice = UcilniceModelMultipleChoiceField(queryset=Ucilnica.objects.objavljene(),
-                                                help_text='Izberite učilnice, ki jih želite rezervirati',
-                                                widget=CheckboxSelectMultiple())
-    use_required_attribute = False
-
-    class Meta:
-        model = Rezervacija
-        fields = ['ucilnice', 'osebe', 'dan', 'dan_konca', 'od', 'do', 'opomba']
-        widgets = {
-            'dan': DateInput(attrs={'placeholder': 'npr. 15. 1. 2019', 'class': 'datepicker'}),
-            'dan_konca': DateInput(attrs={'placeholder': 'ponavadi prazno, lahko tudi npr. 17. 1. 2019', 'class': 'datepicker'}),
-            'opomba': TextInput(attrs={'placeholder': 'npr. izpit Analiza 1 FIN'}),
-        }
-
-    def __init__(self, *args, **kwargs):
-        dovoli_prekrivanja = kwargs.pop('dovoli_prekrivanja', None)
-        super(RezevacijeForm, self).__init__(*args, **kwargs)
-        if dovoli_prekrivanja:
-            self.fields['ignoriraj_prekrivanja'] = BooleanField(
-                required=False, initial=False,
-                help_text='Zavedam se prekrivanj in vseeno želim rezervirati izbrane učilnice '
-                          'v izbranih dnevih in urah.')
-
-    def clean_dan(self):
-        dan = self.cleaned_data['dan']
-        if dan < datetime.date.today():
-            raise ValidationError("Datum rezervacije mora biti v prihodnosti.")
-        return dan
-
-    def clean(self):
-        cleaned = super().clean()
-        if self.errors:
-            return cleaned
-
-        od = cleaned.get('od')
-        do = cleaned.get('do')
-        if od >= do:
-            self.add_error(None, ValidationError("Ura začetka rezervacije mora biti pred uro konca rezervacije."))
-
-        dan = cleaned.get('dan')
-        konec = cleaned.get('dan_konca')
-        if konec:
-            if dan == konec:
-                self.cleaned_data['dan_konca'] = None
-            elif dan > konec:
-                self.add_error(None, ValidationError("Dan začetka rezervacije moda biti "
-                                                     "pred dnevom konca rezervacije."))
-        ignoriraj = cleaned.get('ignoriraj_prekrivanja')
-        if not ignoriraj:
-            self._preveri_konflikte(cleaned)
-
-        return self.cleaned_data
-
-    def _preveri_konflikte(self, cleaned):
-        from urnik.iskanik_konfliktov import IskalnikKonfliktov
-        ucilnice = cleaned.get('ucilnice')
-        dan = cleaned.get('dan')
-        konec = cleaned.get('dan_konca') or dan
-        od = cleaned.get('od')
-        do = cleaned.get('do')
-        iskalnik = IskalnikKonfliktov(ucilnice, dan, konec)
-        iskalnik.dodaj_srecanja()
-        iskalnik.dodaj_rezervacije(Rezervacija.objects.prihajajoce())
-
-        date_format = lambda d: defaultfilters.date(d, "D, j. b")
-        for u in ucilnice:
-            d = dan
-            while d <= konec:
-                konflikti = iskalnik.konflikti(u, d, od, do)
-                for r in konflikti.rezervacije:
-                    oseba = r.osebe.all()[:1]
-                    self.add_error(None, ValidationError(
-                        'Vaša rezervacija bi se prekrivala z rezervacijo osebe %(oseba)s %(dan)s '
-                        'od %(od)i do %(do)i z razlogom %(razlog)s.',
-                        params={
-                            'oseba': oseba[0] if oseba else 'neznan',
-                            'dan': "od {} do {}".format(date_format(r.dan), date_format(r.dan_konca))
-                            if r.dan_konca else "dne {}".format(date_format(r.dan)),
-                            'razlog': r.opomba,
-                            'od': r.od,
-                            'do': r.do,
-                        },
-                        code=RezevacijeForm.PREKRIVANJA,
-                    ))
-
-                for s in konflikti.srecanja:
-                    self.add_error(None, ValidationError(
-                        'Vaša rezervacija bi se prekrivala s predmetom %(predmet)s (%(semester)s), ki se izvaja '
-                        'ob %(dan_v_tednu)s od %(od)i do %(do)i.',
-                        params={
-                            'predmet': s.predmet,
-                            'semester': s.semester,
-                            'dan_v_tednu': dan_tozilnik_mnozina(s.dan),
-                            'od': s.od,
-                            'do': s.do,
-                        },
-                        code=RezevacijeForm.PREKRIVANJA,
-                    ))
-
-                d += datetime.timedelta(days=1)
