+++ urnik/templates/help.html
@@ -22,9 +22,16 @@
       pa sporočite <a href="{% url 'bugreport' %}">tehnično napako</a>.</p>
 
     <h5>Kombiniran pogled na urnik</h5>
-    Če želite hkrati pogledati urnike za več oseb, učilnic in smeri, v meniju ali na prvi strani izberite
-    <code><i class="material-icons left">schedule</i>Kombiniran pogled</code>. Odkljukajte kombinacijo,
-    ki si jo želite ogledati in kliknite na <code>Prikaži kombiniran pogled</code>.
+    <p>
+      Če želite hkrati pogledati urnike za več oseb, učilnic in smeri, v meniju ali na prvi strani izberite
+      <code><i class="material-icons left">schedule</i>Kombiniran pogled</code>. Odkljukajte kombinacijo,
+      ki si jo želite ogledati in kliknite na <code>Prikaži kombiniran pogled</code>. Ker je oseb in predmetov veliko,
+      so vse možnosti naštete v spustnih menijih. V teh menijih lahko izberete več možnosti hkrati. Po njih se lahko
+      premikate s koleščkom na miški, vendar je verjetno hitreje, če natipkate prvih nekaj črk priimka, pa se bo izbor
+      sam premaknil navzdol. Nato lahko izbor premaknete tudi s puščicami na tipkovnici in potrdite s tipko
+      <code>Enter</code>. Meni se nato zapre, če pa želite izbrati še eno možnost, lahko zopet pritisnete
+      <code>Enter</code>, da se meni ponovno odpre in postopek ponovite. Enako velja tudi za predmete (in vse druge
+      spustne menije na strani, npr. pri rezervacijah).</p>
 
     <h5>Tiskanje</h5>
     <p>
